Server side
Crawlers which were built using the Scrapy framework in Python. The crawlers extract this information from the Courses Schedule System, TAU calendar, and the Staff Phone Directory, which are all part of the TAU website. They parse the HTTP pages and extract the relevant data for the app and then update all the collections in the MongoDB.
Also this folder contains cripts for mongoDB.